+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Choice to pre-activate owl statues.
 - Crates shuffle, for OoT and MM.
 - The boat cruise is now faster if you are not playing the archery.
+- New entrance option for Pirate Fortress that includes all the currently shuffled entrances except the Sewers door, considered Overworld.
 
 ### Changed